WIP: separate out databroker stuff
