+++ pytest.ini
@@ -2,3 +2,4 @@
 env =
     D:IS_TEST=1
     XPDAN_SETUP=1
+xfail_strict=true
