+++ shed/tests/test_streams_misc.py
@@ -16,11 +16,11 @@
 # can test individual nodes without having to test everything
 
 import pytest
-from numpy.testing import assert_equal
-from streamz.core import Stream
-
 import shed.event_streams as es
 from shed.tests.utils import SinkAssertion
+from shed.utils import to_event_model
+from streamz.core import Stream
+
 from ..event_streams import star
 
 
@@ -34,7 +34,7 @@ def test_clear():
     assert s.md == {'hello': 'world'}
 
 
-def test_bundle(exp_db, start_uid1, start_uid3):
+def test_bundle():
     source = Stream()
     source2 = Stream()
 
@@ -42,18 +42,21 @@ def test_bundle(exp_db, start_uid1, start_uid3):
     s.sink(star(SinkAssertion(False)))
     L = s.sink_to_list()
 
-    ih1 = exp_db[start_uid1]
-    ih2 = exp_db[start_uid3]
-    s1 = list(exp_db.restream(ih1))
-    s2 = list(exp_db.restream(ih2))
+    s1 = list(to_event_model(
+        [_ for _ in range(3)],
+        output_info=[('pe1_image', {'dtype': 'array'})]
+    ))
+    s2 = list(to_event_model(
+        [_ for _ in range(3)],
+        output_info=[('pe1_image', {'dtype': 'array'})]
+    ))
     uids = set([doc['uid'] for name, doc in s1] + [doc['uid'] for name, doc in
                                                    s2])
     for b in s2:
         source2.emit(b)
     for a in s1:
         source.emit(a)
-    assert len(L) == len(list(exp_db.get_events(ih1))) + len(
-        list(exp_db.get_events(ih2))) + 3
+    assert len(L) == len(s1) + len(s2) - 3
 
     assert_docs = set()
     for l in L:
@@ -63,19 +66,23 @@ def test_bundle(exp_db, start_uid1, start_uid3):
         assert n in assert_docs
 
 
-def test_combine_latest(exp_db, start_uid1, start_uid3):
+def test_combine_latest():
     source = Stream()
     source2 = Stream()
 
     L = es.combine_latest(source, source2, emit_on=source).sink_to_list()
-    ih1 = exp_db[start_uid1]
-    ih2 = exp_db[start_uid3]
-    s = exp_db.restream(ih1)
-    s2 = exp_db.restream(ih2)
+    s1 = list(to_event_model(
+        [_ for _ in range(3)],
+        output_info=[('pe1_image', {'dtype': 'array'})]
+    ))
+    s2 = list(to_event_model(
+        [_ for _ in range(2)],
+        output_info=[('pe1_image', {'dtype': 'array'})]
+    ))
     for b in s2:
         source2.emit(b)
         print(b)
-    for a in s:
+    for a in s1:
         source.emit(a)
 
     assert_docs = set()
@@ -84,205 +91,6 @@ def test_combine_latest(exp_db, start_uid1, start_uid3):
         assert l1 != l2
         if name == 'event':
             assert l2['seq_num'] == 2
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-def test_query(exp_db, start_uid1):
-    source = es.EventStream()
-
-    def qf(db, docs):
-        return db(uid=docs[0]['uid'])
-
-    hdr = exp_db[start_uid1]
-    s = hdr.documents()
-
-    dp = es.Query(exp_db, source, qf,
-                  query_decider=lambda x, y: [next(iter(x))])
-    L = dp.sink_to_list()
-
-    dp2 = es.QueryUnpacker(exp_db, dp)
-    L2 = dp2.sink_to_list()
-
-    for a in s:
-        source.emit(a)
-
-    assert_docs = set()
-    for l in L:
-        assert_docs.add(l[0])
-        if l[0] == 'event':
-            assert l[1]['data']['hdr_uid'] == start_uid1
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-    assert_docs = set()
-    for l, ll in zip(L2, hdr.documents()):
-        assert_docs.add(l[0])
-        assert l[0] == ll[0]
-        if l[0] is 'start':
-            assert l[1] == ll[1]
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-def test_empty_query(exp_db, start_uid1):
-    source = es.EventStream()
-
-    def qf(db, docs):
-        return db(hello='world')
-
-    def qd(res, docs):
-        try:
-            rv = [next(iter(res))]
-        except StopIteration:
-            rv = []
-        return rv
-
-    hdr = exp_db[start_uid1]
-    s = hdr.documents()
-
-    dp = es.Query(exp_db, source, qf,
-                  query_decider=qd)
-    L = dp.sink_to_list()
-
-    dp2 = es.QueryUnpacker(exp_db, dp)
-    L2 = dp2.sink_to_list()
-
-    for a in s:
-        source.emit(a)
-
-    assert_docs = set()
-    for l in L:
-        assert_docs.add(l[0])
-        if l[0] == 'start':
-            assert l[1]['n_hdrs'] == 0
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-    for n in ['start', 'descriptor', 'stop']:
-        assert n in assert_docs
-    assert len(L2) == 0
-
-
-def test_query_many_headers(exp_db):
-    source = es.EventStream()
-
-    def qf(db, docs):
-        return db(sc_dk_field_uid={'$exists': True})
-
-    s = [('start', None)]
-
-    dp = es.Query(exp_db, source, qf)
-    L = dp.sink_to_list()
-
-    dp2 = es.QueryUnpacker(exp_db, dp)
-    dp2.sink(print)
-    L2 = dp2.sink_to_list()
-
-    for a in s:
-        source.emit(a)
-
-    assert len(L) == 6
-    assert_docs = set()
-    for l in L:
-        assert_docs.add(l[0])
-        if l[0] == 'start':
-            assert l[1]['n_hdrs'] == 3
-        if l[0] == 'event':
-            assert l[1]['data']['hdr_uid'] in list(d['start']['uid'] for d in
-                                                   qf(exp_db, 'hi'))
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-    assert_docs = set()
-    assert len(L2) == 3 * 3 + 5 + 5 + 2
-    for l in L2:
-        assert_docs.add(l[0])
-        assert l[0]
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-@pytest.mark.xfail(raises=RuntimeError)
-def test_query_too_many_headers(exp_db):
-    source = es.EventStream()
-
-    def qf(db, docs):
-        return db(sc_dk_field_uid={'$exists': True})
-
-    s = [('start', None)]
-
-    es.Query(exp_db, source, qf, max_n_hdrs=1)
-
-    for a in s:
-        source.emit(a)
-
-
-def test_bundle_single_stream(exp_db):
-    source = es.EventStream()
-
-    def qf(db, docs):
-        return db(sc_dk_field_uid={'$exists': True})
-
-    s = [('start', {})]
-
-    dp = es.Query(exp_db, source, qf)
-
-    dp2 = es.QueryUnpacker(exp_db, dp)
-
-    dpf = es.BundleSingleStream(dp2, dp)
-
-    L = dpf.sink_to_list()
-    dpf.sink(print)
-
-    for a in s:
-        source.emit(a)
-
-    assert_docs = set()
-    assert len(L) == 3 + 5 + 5 + 2
-    for l in L:
-        assert_docs.add(l[0])
-        assert l[0]
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-def test_bundle_single_stream_control_int(exp_db):
-    source = es.EventStream()
-
-    def qf(db, docs):
-        return db(sc_dk_field_uid={'$exists': True})
-
-    s = [('start', {})]
-
-    dp = es.Query(exp_db, source, qf)
-
-    dp2 = es.QueryUnpacker(exp_db, dp)
-
-    dpf = es.BundleSingleStream(dp2, 2)
-
-    L = dpf.sink_to_list()
-    dpf.sink(print)
-
-    for a in s:
-        source.emit(a)
-
-    assert_docs = set()
-    assert len(L) == 3 + 5 + 2
-    for l in L:
-        assert_docs.add(l[0])
-        assert l[0]
-        if l[0] == 'stop':
-            assert l[1]['exit_status'] == 'success'
     for n in ['start', 'descriptor', 'event', 'stop']:
         assert n in assert_docs
 
@@ -327,7 +135,7 @@ def test_workflow(exp_db, start_uid1):
 
 
 @pytest.mark.xfail(raises=ValueError)
-def test_event_contents_fail(exp_db, start_uid1):
+def test_event_contents_fail():
     source = Stream()
 
     from operator import add
@@ -341,8 +149,10 @@ def test_event_contents_fail(exp_db, start_uid1):
                 output_info=oi)
     dp.sink(star(SinkAssertion(True)))
 
-    ih1 = exp_db[start_uid1]
-    s = exp_db.restream(ih1, fill=True)
+    s = list(to_event_model(
+        [_ for _ in range(2)],
+        output_info=[('pe1_image', {'dtype': 'array'})]
+    ))
     for a in s:
         source.emit(a)
 
@@ -468,7 +278,7 @@ def test_string_workflow(exp_db, start_uid1):
         assert n in assert_docs
 
 
-def test_split(exp_db, start_uid1):
+def test_split():
     source1 = Stream()
     source2 = Stream()
 
@@ -477,32 +287,22 @@ def test_split(exp_db, start_uid1):
     L1 = uz.split_streams[0].sink_to_list()
     L2 = uz.split_streams[1].sink_to_list()
 
-    h1 = exp_db[start_uid1]
+    s = list(to_event_model(
+        [_ for _ in range(3)],
+        output_info=[('pe1_image', {'dtype': 'array'})]
+    ))
 
-    for s1, s2 in zip(h1.documents(), h1.documents()):
+    for s1, s2 in zip(s, s):
         source1.emit(s1)
         source2.emit(s2)
 
     assert_docs = set()
-    for s, s1, s2 in zip(h1.documents(), L1, L2):
+    for s, s1, s2 in zip(s, L1, L2):
         assert s == s1
         assert s == s2
         assert_docs.add(s1[0])
     for n in ['start', 'descriptor', 'event', 'stop']:
         assert n in assert_docs
-
-
-def test_fill_events(exp_db, start_uid1):
-    source = Stream()
-    dp = es.fill_events(exp_db, source)
-    L = dp.sink_to_list()
-
-    h1 = exp_db[start_uid1]
-    for s in h1.documents():
-        source.emit(s)
-
-    for a, b in zip(L, h1.documents(fill=True)):
-        assert_equal(a[1], b[1])
 
 
 def test_descriptor_no_output_info(exp_db, start_uid1):
@@ -528,3 +328,37 @@ def test_descriptor_no_output_info(exp_db, start_uid1):
         assert_docs.add(l[0])
     for n in ['start', 'descriptor', 'event', 'stop']:
         assert n in assert_docs
+
+
+@pytest.mark.xfail(raises=RuntimeError,
+                   reason="Received RunStop before RunStart.")
+def test_stop_before_start():
+    s = list(to_event_model(
+        [_ for _ in range(3)],
+        output_info=[('pe1_image', {'dtype': 'array'})]
+    ))
+    source = Stream()
+    from operator import add
+    es.map(add, source, 1, input_info={0: 'pe1_image'})
+
+    s = [s[-1]]
+    for _ in range(2):
+        for d in s:
+            source.emit(d)
+
+
+@pytest.mark.xfail(raises=RuntimeError,
+                   reason="Received EventDescriptor before RunStart")
+def test_event_before_start():
+    s = list(to_event_model(
+        [_ for _ in range(3)],
+        output_info=[('pe1_image', {'dtype': 'array'})]
+    ))
+    source = Stream()
+    from operator import add
+    es.map(add, source, 1, input_info={0: 'pe1_image'})
+
+    s = [s[1]]
+    for _ in range(2):
+        for d in s:
+            source.emit(d)
